Separate out the dictionary service
